+++ CHANGES.md
@@ -10,6 +10,7 @@ Improvements 🙌:
  - Fix autocompletion issues and add support for rooms and groups
  - Introduce developer mode in the settings (#796)
  - Improve devices list screen
+ - Add settings for rageshake sensibility
 
 Other changes:
  -